+++ R/proc_reg.R
@@ -195,6 +195,15 @@ proc_reg <- function(data,
   #                    error = function(cond) {opaired})
 
   # Parameter checks
+
+  if (!"data.frame" %in% class(data)) {
+    stop("Input data is not a data frame.")
+  }
+
+  if (nrow(data) == 0) {
+    stop("Input data has no rows.")
+  }
+
   nms <- names(data)
 
   if (length(model) == 0) {