+++ R/proc_ttest.R
@@ -364,6 +364,14 @@ proc_ttest <- function(data,
                      error = function(cond) {opaired})
 
   # Parameter checks
+  if (!"data.frame" %in% class(data)) {
+    stop("Input data is not a data frame.")
+  }
+
+  if (nrow(data) == 0) {
+    stop("Input data has no rows.")
+  }
+
   nms <- names(data)
 
   if (!is.null(by)) {